让长方形计算使用 double

+++ rectangle/main.c
@@ -1,27 +1,27 @@
 #include <stdio.h>
-int length , width , perimeter , square;
+double length , width , perimeter , square;
 
-int compute_perimeter(int length,int width);
-int compute_square(int length, int width);
+double compute_perimeter(double length, double width);
+double compute_square(double length, double width);
 
 int main()
 {
 
     printf("Please input the length:");
-    scanf("%d", &length);
+    scanf("%lf", &length);
     printf("Please input the width:");
-    scanf("%d", &width);
-    printf("perimeter is %d\n", compute_perimeter(length, width));
-    printf("square is %d", compute_square(length, width));
+    scanf("%lf", &width);
+    printf("perimeter is %lf\n", compute_perimeter(length, width));
+    printf("square is %lf", compute_square(length, width));
     return 0;
 }
 
-int compute_perimeter(int length, int width)
+double compute_perimeter(double length, double width)
 {
     return length + length + width + width;
 }
 
-int compute_square(int length, int width)
+double compute_square(double length, double width)
 {
     return length * width;
 }
